class Map FileIO
Implemented fileIO for the map class.  Because I havent written in python in over like a week, I have been extremely slow with my implementation with simplty reading and writing into and out of a file (actually i havent even written out).... I can't believe that I am so slow!  But that's okay slow and steady wins the race.  And I will not compare myself to my peers, even though they might be faster and more intelligent than I am with programming.  I believe that half of the software circulating in our economy simply exist to absorb economic incentives and are not used to try to make people's health or well-being more benefitical.  (just look at cultural anthrolopolgical views on the impact of technology on socities).